add neo scripts + update paleo scripts

+++ paleo_pseudogenes/collinear_block_identification/get_genelists.sh
@@ -9,9 +9,11 @@ plaza_name=$2 # species abbreviation used in PLAZA, e.g. ath
 wd="/scratch/recent_wgds/paleo_polyploids/collinearity"
 gff="/scratch/recent_wgds/data/$species/gff/annotation.selected_transcript.all_features.${plaza_name}.gff3"
 # after low filtering
-pseudogenes="/scratch/recent_wgds/paleo_polyploids/filtering/${species}_filtered_low_pgenes_unprocessed.txt"
+#pseudogenes="/scratch/recent_wgds/paleo_polyploids/filtering/${species}_filtered_low_pgenes_unprocessed.txt"
+# after filtering out pseudogenes that reside in introns
+pseudogenes="/scratch/recent_wgds/paleo_polyploids/collinearity/filtered_no_intron/${species}_filtered_no_intronic.txt"
 
-analysis="i-adhore/with_middle_filtering_level2_false/${species}"
+analysis="i-adhore/no_introns2/${species}"
 
 ## Create genome lists
 mkdir $analysis/genome_lists
@@ -24,10 +26,13 @@ awk -F "\t" '{print $1"\t"$2"\t"$3"\t"$4"\t"$15}' $pseudogenes > pgenes.txt
 ### Combine processed files and create a file per chromosome/scaffold/...
 cat genes.txt pgenes.txt | sort -k1,1 -k2,2n | awk '{print $1"\t"$5$4}' | awk -F "\t" '{print $2>$1}'
 rm genes.txt pgenes.txt
-### If chromosome level assembly remove scaffolds, otherwise comment following line out (e.g. Amborella):
-rm *scaf*
-rm *random*
-rm *super*
-rm chrUn
-rm ChrM
-rm ChrC
+### If chromosome level assembly remove scaffolds, except for Amborella:
+if [[ $species != "Amborella_trichopoda" ]]
+then
+    rm *scaf*
+    rm *random*
+    rm *super*
+    rm chrUn
+    rm ChrM
+    rm ChrC
+fi
